Day 03

+++ src/days/day_01.rs
@@ -1,7 +1,7 @@
-pub fn part_01(input: &String) -> i32 {
-	let expenses: Vec<i32> = input
+pub fn part_01(input: &String) -> u64 {
+	let expenses: Vec<u64> = input
 		.split('\n')
-		.map(|e| e.parse::<i32>().expect("Failed to parse input"))
+		.map(|e| e.parse::<u64>().expect("Failed to parse input"))
 		.collect();
 	for i in (1..expenses.len()).rev() {
 		for j in 0..i - 1 {
@@ -14,10 +14,10 @@ pub fn part_01(input: &String) -> i32 {
 	panic!("This is not good...")
 }
 
-pub fn part_02(input: &String) -> i32 {
-	let expenses: Vec<i32> = input
+pub fn part_02(input: &String) -> u64 {
+	let expenses: Vec<u64> = input
 		.split('\n')
-		.map(|e| e.parse::<i32>().expect("Failed to parse input"))
+		.map(|e| e.parse::<u64>().expect("Failed to parse input"))
 		.collect();
 	for i in (2..expenses.len()).rev() {
 		for j in (1..i - 1).rev() {
